fix:otp

+++ app/Services/OtpService.php
@@ -41,27 +41,27 @@ class OtpService
 
     static function verifyOtp($mobile,$code)
     {
-        $expireTime = now()->timezone('Asia/Tehran');
+        $expireTime = now();
         $mobile = Helpers::mobileNumberNormalize($mobile);
         $user = User::query()->where('mobile',$mobile)->first();
-        $otpGeneratedCode = Otp::query()
-            ->where('user_id',$user->id)
-            ->where('expired_at','>=',$expireTime);
-        if(!$otpGeneratedCode->exists()){
-            if($otpGeneratedCode->withTrashed()->exists()){
-                return 'deleted';
-            }
-            return 'expired';
-        } else {
-            if($otpGeneratedCode->first()->code==$code){
-                $otpGeneratedCode->delete();
-                if(!$user->hasVerifiedMobile())
-                    $user->markMobileAsVerified();
+        if(is_null($user->mobile_verified_at)) {
+            $otpGeneratedCode = Otp::query()
+                ->where('user_id', $user->id)
+                ->where('expired_at', '>=', $expireTime);
+            if (!$otpGeneratedCode->exists()) {
+                return false;
+            } else {
+                if ($otpGeneratedCode->first()->code == $code) {
+                    $otpGeneratedCode->delete();
+                    if (!$user->hasVerifiedMobile())
+                        $user->markMobileAsVerified();
 
-                return 'verified';
+                    return true;
+                }
             }
-        }
-        return false;
+            return false;
+        } else
+            return true;
     }
 
 }
